move data files to api folder
