+++ pages/api/credit-card/CreditCardWordProblemData.ts
