+++ pages/api/credit-card/DebitCardWordProblemData.ts
